skills/el-monstruo-armero/templates/sprint1/requirements.txt
# El Monstruo — Sprint 1 Requirements
# Generado: 2026-04-09
# IMPORTANTE: ANCLAR versiones críticas

# === KERNEL ===
langgraph==1.1.6
langchain-core>=0.3.0
langgraph-checkpoint-postgres>=0.1.0

# === ROUTER ===
# ALERTA: NUNCA usar >= para litellm. CVE-2026-35030
litellm==1.83.3

# === MEMORIA ===
mem0ai>=0.1.0
supabase>=2.0.0

# === OBSERVABILIDAD ===
langfuse>=2.0.0

# === LLM PROVIDERS ===
openai>=1.50.0
anthropic>=0.30.0
google-genai>=1.0.0

# === WEB FRAMEWORK ===
fastapi>=0.115.0
uvicorn>=0.30.0
python-telegram-bot>=21.0

# === UTILIDADES ===
requests>=2.32.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# === WEB SCRAPING (opcional Sprint 1) ===
# firecrawl-py>=0.1.0

# === AUDIO (opcional) ===
# elevenlabs>=1.0.0
